requirements.txt
streamlit
pandas
plotly
u8darts[all]
torch
notionhelper
seaborn
holidays